moved RmrkFunc to types crate
