+++ chainextensions/rmrk/impl/Cargo.toml
@@ -25,8 +25,10 @@ log = "0.4.16"
 
 # RMRK
 chain-extension-traits = { path = "../../traits", default-features = false }
+rmrk-chain-extension-types = { path = "../types", default-features = false }
 pallet-rmrk-core = { git = "https://github.com/AstarNetwork/rmrk-substrate", branch = "polkadot-v0.9.26", default-features = false }
 rmrk-traits = { version = "0.0.1", git = "https://github.com/AstarNetwork/rmrk-substrate", branch = "polkadot-v0.9.26", default-features = false }
+
 
 [features]
 default = ["std"]
@@ -42,6 +44,7 @@ std = [
     "pallet-rmrk-core/std",
     "pallet-uniques/std",
     "rmrk-traits/std",
+    "rmrk-chain-extension-types/std",
     "scale-info/std",
     "sp-std/std",
     "sp-core/std",
